Engine Base Upload
Cronos Engine base code upload, starting off from physics subject base code. This one has been modifyied to serve the Cronos Engine purposes.

The globals file has been widely extended to have more tools, a premake file has been built in order to set up the project as well as a .bat file to generate it when needed (so you don't have to call a cmd command) departing from that premake file (there is a possible bug here regarding the project's environment, I'm working on it).

All the file structure for the project has been modified to have it all cleaner and more accessible. Also, in the physics base code, the Application* App created was not specifyied as extern, so to access it, you had to include application.h, but that has been changed, now App is actually extern.

Also, a pre-compiled header has been created so we can precompile files that we need and won't change in order to not to compile them in runtime everytime we include them.

A namespace for the Engine elements (namespace "Cronos") has been created to diferentiate Engine's stuff from other, to prevent possible conflicts and to group entities into the same engine's namespace.

Added timers too an dt calculations as well as a plane as a wirefrom and with the axis active to render it and have a reference and a visual starting point to do things.

Finally, implemented a way (not tested at 100% yet since I need to start with OpenGL to do it) to make a callback to a personalized function when OpenGL crashes (you know, if OGL fails - not with a fatal error or crash- it just doesn't warns, just stops drawing, and with this we can make it say something).