Camera movement update
Just changed the camera moving speed as well as doubling it when LShift is pressed.

Also, I have implemented the movement as requested.

About focus and orbiting, it works, for free-orbiting and focus to center, but to make it object-referenced, we just need to pass the object axis or reference :)