Update on Camera controls
Now with ALT + LMouseButton you can orbit around the center. This is in a function in which we have to pass a reference, so when we need to make it rotate around an object, we just call this function passing the object's reference and we're done :)

I also moved the code that get the mouse movement into a function so it was all more readable and abstract.