Update on Camera controls & Log corrected
Added camera panning with mouse middle button and zoom in/out with mouse mid button scroll. Also, I've changed the way in which camera is moved Up & Down in Y axis because it conflicted with camera focus-to-center.

Also corrected a wrong info given in log when logging the compilation message and ordered the code.