Corrected Camera Zoom
Just make it work between two given FOVs (15 and 90), else won't zoom because passing those FOVs limits is not recommended.

About the code clean, I need to think about how we will do it, because right now the renderer controls the projection matrix, but the camera is kind of this projection matrix and also is an independent module, so this needs to be restructured.