Updated Camera
I have refactored some things and created soome camera methods. Actually, I've also added some things which actually make it be a true camera.

I have changed its name to 'EngineCamera' since it fits better, we are not doing "Cameras" in the engine by now (for that we will need game objects) so, in the meantime, this will be the actual engine camera and the module will act as it.

Also, since the window resize event implied to change the 'engine's view projection', I have made a method in the engine camera which recalculates the camera's view projection matrix. In fact, I have changed the resize function, now is in window module (where should it be, eh, Rick ¬¬') and not in Renderer (wtf renderer should be only to render xdxd), as well, Renderer do not stores anymore the view, model and projection matrix. From now on, model matrices will be in the own models, view and projection will be in engine camera module (and probably, if we add cameras feature, each camera will have their own VP matrix).

So the window, when resizing, calls the camera to recalculate the projection. I have even added an aspect ratio variable, a FOV, near and far planes... So we can calculate all that properly and even set it at our way.