Json Implemented (now yes)
I did it in dev branch, but wasn't correct, so I reverted the commit and did it here.

Also I have made functions in application to load and save a json file and detect if, when loading, there is no file.

Now we need to make a system to serialize and all great :)