Changed Primitives and make them transform
I discovered that the thing I did with buffers specially for primitives  (to make them be of Ushort indices and not Uints) could be avoided with a way of translating the ushort vector of primitive's indices from parshapes into a uint vector, so I changed that.

Also I made a system to transform models (including primitives) --> Now we can move, scale and rotate them.