Changed Models' transform functions and added a couple of new primitives
Just added some special primitives such as disk (this will be to do solid cylinders), a subdivided sphere and a rock creator.

Also I have reordered a bit the primitive class so it's more readable and we can avoid to repeat code.

Finally, I had to change the functions to transform models(done with matrices now) because the transormation was not well performed. Now it seems it does, although it still is a bit unperformative, I think.