Fixed System Info Values &
A little fix on system info return values was needed in order to be properly shown on UI (Just conversions to MB or GB). There is still a problem with MMGR values, but this can be commented in class.

Also, commented some Window functions to set window parameters since they are done from UI and might need to be reworked. This is done just to avoid problems or conflicts.