Minor Corrections and Textures & Shaders Update
·Deleted Stupid Comments and unused code

·Redefined Shaders to handle matrices and textures from there

·Rework on Textures ----------------------------
Now we have textures types, they are created from the textures manager and directly loaded from assimp (still we need to make a workaround for primitives).
They can be changed (untested), the mesh draw function gets a shader, binds it, passes its textures & transformations and draws. Then unbinds it.

I don't know why, but it seems that the texture is not properly loaded with another model (for instance from 3ds max), I'll look into this.
---------------------------------------------------

·Corrected model's planes normals calculation (now they are properly calculated and rendered)

·Corrected (a bit) the model axis calculation, but still is not correct. Needs more work (as well as transforms, but all this must be left behind by now until the 1st delivery).