Solved Shaders/Materials Problems
I had some problems with binding/unbinding the shader in the material components or in the mesh...

Also to render primitives, it gave many problems. Now they are fixed :), the shader is bind from the mesh (so the primitives can be rendered), the model matrices of each object is set from the transform component and the textures from the material one.

All ok :), I just want to commit suicide.