Changed scene floor plane
Now is a plane primitive and not a direct-mode OpenGL plane, which was horrible.

Now I just need to control transformations to put it in place :)

BTW, there might be a problem with cull face with this if it's set to unactive, because the plane render will active once again. It needs a little rework