Update on Transformations
Fixed some things done wrong on transformations. Also, we can add/substract a transformation (like: we have this rotation and add it this one, and it works with positions and scalations too) and set them up (like: we have this position, set it to this one, the same for rot/scale).

The problem is that UI rotation and the scalation (and also the camera, ya que estamos) need a bit more fix, they are not properly performed.