Update on direct mode drawing, code reorganization and comments delete
Managed to focus all direct mode drawing in the Renderer class and to successfully draw all lines/cubes... whatever.
Now I need to make the lines or the vertex/planes normals to be transformed, they are not.

Also, deleted unuseful comments

Reordered a bit the components' code

Fixed wireframe draw mode problems (and color)

Set a default draw color for renderer and a default linewidth