Started materials rework
I just started with all this mess... The material components, as materials perse, have been fully changed... They now store a material (from a material class).

Everything work pretty much the same, but instead of being the component a material, they just reference a material.

By now, when a component material is created, a new material is created with it, but this must be changed, they SHOULD NOT create materials, they should just hold its data.

I'll work on it. BTW, that makes the materials to crash fatally when deleting textures... It is under control in this moment since I think that it is because new materials are created along with each comp.Mat.

Also, I deleted some unuseful comments and commented some of the current texture types for further implementation (so they don't bother me while doing all this shit)