Mouse picking working as expected
I fixed the issue we had, until now, the mouse picking sometimes picked the wrong object because it was "nearer" to the camera since it was bigger, but there was one object better to pick since it was before the one picked.

Now that is fixed, It works now with the distance from camera to intersecting point between object-forward ray, not with the distance from camera to object.